dashboard_control: carve dispatch.rs (pure move)
